Extract integration links to Tune.Link
