+++ lib/tune_web/templates/album/show.html.eex
@@ -23,11 +23,11 @@
             <%= content_tag :span, total_duration(@album), class: "total-duration" %>
           <% end %>
           <div class="reference-links">
-            <%= link to: last_fm_link(@album), target: "_blank", class: "svg-link last-fm" do %>
+            <%= link to: Link.last_fm(@album), target: "_blank", class: "svg-link last-fm" do %>
               <%= render TuneWeb.TrackView, "icon_last_fm.html", [] %>
               <%= content_tag :span, gettext("Open in Last.fm"), class: "visually-hidden" %>
             <% end %>
-            <%= link to: youtube_link(@album), target: "_blank", class: "svg-link youtube" do %>
+            <%= link to: Link.youtube(@album), target: "_blank", class: "svg-link youtube" do %>
               <%= render TuneWeb.TrackView, "icon_youtube.html", [] %>
               <%= content_tag :span, gettext("Open in YouTube"), class: "visually-hidden" %>
             <% end %>
