+++ lib/tune_web/templates/track/list.html.eex
@@ -5,13 +5,13 @@
         <span class="track-number">&#127926;</span>
         <span class="name" phx-click="play" phx-value-uri="<%= track.uri %>" phx-value-context-uri="<%= @album.uri %>"><%= track.name %></span>
         <span class="duration"><%= Tune.Duration.hms(track.duration_ms) %></span>
-        <%= link to: last_fm_link(track, @album, @artist), class: "last-fm", target: "_blank" do %>
+        <%= link to: Link.last_fm(track, @album, @artist), class: "last-fm", target: "_blank" do %>
           <%= render "icon_last_fm.html", [] %>
         <% end %>
-        <%= link to: youtube_link(track, @artist), class: "youtube", target: "_blank" do %>
+        <%= link to: Link.youtube(track, @artist), class: "youtube", target: "_blank" do %>
           <%= render "icon_youtube.html", [] %>
         <% end %>
-        <%= link to: musixmatch_link(track, @artist), class: "lyrics", target: "_blank" do %>
+        <%= link to: Link.musixmatch(track, @artist), class: "lyrics", target: "_blank" do %>
           <%= render "icon_lyrics.html", [] %>
         <% end %>
       </li>
@@ -20,13 +20,13 @@
         <span class="track-number"><%= track.track_number %></span>
         <span class="name" phx-click="play" phx-value-uri="<%= track.uri %>" phx-value-context-uri="<%= @album.uri %>"><%= track.name %></span>
         <span class="duration"><%= Tune.Duration.hms(track.duration_ms) %></span>
-        <%= link to: last_fm_link(track, @album, @artist), class: "last-fm", target: "_blank" do %>
+        <%= link to: Link.last_fm(track, @album, @artist), class: "last-fm", target: "_blank" do %>
           <%= render "icon_last_fm.html", [] %>
         <% end %>
-        <%= link to: youtube_link(track, @artist), class: "youtube", target: "_blank" do %>
+        <%= link to: Link.youtube(track, @artist), class: "youtube", target: "_blank" do %>
           <%= render "icon_youtube.html", [] %>
         <% end %>
-        <%= link to: musixmatch_link(track, @artist), class: "lyrics", target: "_blank" do %>
+        <%= link to: Link.musixmatch(track, @artist), class: "lyrics", target: "_blank" do %>
           <%= render "icon_lyrics.html", [] %>
         <% end %>
       </li>
