+++ lib/tune_web/views/album_view.ex
@@ -2,7 +2,7 @@ defmodule TuneWeb.AlbumView do
   @moduledoc false
   use TuneWeb, :view
 
-  alias Tune.Spotify.Schema.Album
+  alias Tune.{Link, Spotify.Schema.Album}
   alias TuneWeb.TrackView
 
   @default_artwork "https://via.placeholder.com/300"
@@ -34,30 +34,5 @@ defmodule TuneWeb.AlbumView do
       tracks_count,
       %{formatted_duration: formatted_duration}
     )
-  end
-
-  @spec last_fm_link(Album.t()) :: String.t()
-  defp last_fm_link(album) do
-    artist_name =
-      album
-      |> Album.main_artist()
-      |> Map.fetch!(:name)
-
-    Path.join([
-      "https://www.last.fm/music",
-      URI.encode(artist_name),
-      URI.encode(album.name)
-    ])
-  end
-
-  @spec youtube_link(Album.t()) :: String.t()
-  defp youtube_link(album) do
-    artist_name =
-      album
-      |> Album.main_artist()
-      |> Map.fetch!(:name)
-
-    q = [search_query: artist_name <> " " <> album.name]
-    "https://www.youtube.com/results?" <> URI.encode_query(q)
   end
 end
